Add more directories to the gitignore
These directories are generated by the gulp/webpack process. I don’t
want anything in the repo that will be generated, which is why they’ve
been added. It is a little odd to have the public directory in there,
but that seems to be the output for the app being “built”.