Change hands reducer test to catch an edge case
Previously, I was doing this to know whether to remove a card from the
hand or not:

```js
return item.face !== card.face && item.suit !== card.suit;
```

This had a bug, though, because of the `&&` instead of `||`. We want to
filter out items that are the card we’re looking for. Because of the
AND, though, we were stopping after the `face` check and filtering out
any cards that were the same face as the card we were looking for.

This test would’ve caught that edge case.

Also, fixed the logic by going with this, instead:

```js
return !(item.face === card.face && item.suit === card.suit);
```