Use different card svgs and don't use a sprite sheet
The new card svgs are smaller (in kb). Also, we’re not using a sprite
sheet anymore. Initially, this was because I couldn’t figure out how to
 et the new svgs to work correctly when combined into a sprite sheet.
Technically it’s inefficient to load 34 different images on demand, but
in practice I don’t notice any performance issues. If anything, the
images show up faster, because the old sprite sheet was several
megabytes.