Avoid unnecessary object lookup and use the existing state
Seems kind of silly to lookup the right character from the object when we already have 'this.state'. Having said that, the downside if what we're doing now is that we have to have the "magic" strings of "hearts" and "diamonds". Considering that those are unlikely to change, though, it doesn't seem like a big deal.